+++ packages/models-library/src/models_library/api_schemas_clusters_keeper/ec2_instances.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from pydantic import ByteSize, PositiveInt
+
+
+@dataclass(frozen=True)
+class EC2InstanceType:
+    name: str
+    cpus: PositiveInt
+    ram: ByteSize
